# `<bit>`: Use popcount for `has_single_bit()` if it's always available