--- a/stl/inc/bit
+++ b/stl/inc/bit
@@ -84,14 +84,14 @@ _NODISCARD constexpr int countl_zero(_Ty _Val) noexcept;
 
 _EXPORT_STD template <_Standard_unsigned_integral _Ty>
 _NODISCARD constexpr bool has_single_bit(const _Ty _Val) noexcept {
-#if defined(_POPCNT_INTRINSICS_ALWAYS_AVAILABLE) && _POPCNT_INTRINSICS_ALWAYS_AVAILABLE
+#if _POPCNT_INTRINSICS_ALWAYS_AVAILABLE
     if (_STD is_constant_evaluated()) {
         return (_Val ^ (_Val - 1)) > _Val - 1;
     }
     return _Unchecked_popcount(_Val) == 1;
-#else // ^^^ popcnt intrinsic always available / popcnt intrinsic not always available vvv
+#else // ^^^ _POPCNT_INTRINSICS_ALWAYS_AVAILABLE / !_POPCNT_INTRINSICS_ALWAYS_AVAILABLE vvv
     return (_Val ^ (_Val - 1)) > _Val - 1;
-#endif // ^^^ popcnt intrinsic not always available ^^^
+#endif // ^^^ !_POPCNT_INTRINSICS_ALWAYS_AVAILABLE ^^^
 }
 
 inline void _Precondition_violation_in_bit_ceil() noexcept {}