@AlexGuteniev 2025-05-23
> @AlexGuteniev, do I need to update benchmarks, or simply put benchmark numbers as part of the commit message?

I think the current benchmarks are fine, just post before/after number