@AlexGuteniev 2025-05-23
Ideally, we'd need ARM64 numbers as well. But I think we can do without them.