@AlexGuteniev 2025-05-23
> I think the current benchmarks are fine, just post before/after number

You also need to pass somehow `/arch:AVX` or greater to make it different, perhaps by temporarily alterint `CmakeFIle.txt` locally, or by any other working method.