@pps83 2025-05-23
> You also need to pass somehow `/arch:AVX` or greater to make it different, perhaps by temporarily alterint `CmakeFIle.txt` locally, or by any other working method.

shouldn't unit tests automatically run with/without these switches?