@AlexGuteniev 2025-05-23
> shouldn't unit tests automatically run with/without these switches?

Unit tests should. 
I mean benchmarks, which currently don't