@pps83 2025-05-23
I'm getting conflicting results for the test: popcount on my pc is somehow slower by around 30%
it's a tiger-lake-u core i7-11390H, shouldn't have slow popcount :)