@pps83 2025-05-23
> I'm getting conflicting results for the test: popcount on my pc is somehow slower by around 30% it's a tiger-lake-u core i7-11390H, shouldn't have slow popcount :)

no idea how is that it doesn't get inlined :)

![image](https://github.com/user-attachments/assets/a61f61d2-041f-4427-8809-1ebc866d5379)