@pps83 2025-05-23
popcount is on the right. So, it does improve, but will be 30% slower for projects using /Ob1
![Uploading image.png…]()