@AlexGuteniev 2025-05-23
> Imo, all these functions should be marked inline.

No; I asked a while ago, see #619