@pps83 2025-05-23
> So you probably mean 1.05 to 1.1 times speed up

updated. I based it on my previous commit message: https://github.com/microsoft/STL/pull/5367

(actual git history doesn't have that anyways)