@AlexGuteniev 2025-05-23
Mixed results for `_if` version, but consistent improvement of the branchless one

i5-1235U (Alder Lake)

P Cores

Benchmark                      | Before   |  After   | Speedup
-------------------------------|----------|----------|---------
bm_has_single_bit_if<uint8_t>  | 9.78 ns  |  9.90 ns |   0.99
bm_has_single_bit_if<uint16_t> | 12.1 ns  |  10.1 ns |   1.20
bm_has_single_bit_if<uint32_t> | 11.4 ns  |  12.3 ns |   0.93
bm_has_single_bit_if<uint64_t> | 9.66 ns  |  12.7 ns |   0.76
bm_has_single_bit<uint8_t>     | 4.44 ns  |  5.56 ns |   0.80
bm_has_single_bit<uint16_t>    | 4.66 ns  |  3.59 ns |   1.30
bm_has_single_bit<uint32_t>    | 4.66 ns  |  3.59 ns |   1.30
bm_has_single_bit<uint64_t>    | 4.90 ns  |  3.47 ns |   1.41

E Cores

Benchmark                      |    Time  |    Time | Speedup
-------------------------------|----------|---------|---------
bm_has_single_bit_if<uint8_t>  | 14.7 ns  | 14.0 ns |  1.05
bm_has_single_bit_if<uint16_t> | 16.6 ns  | 14.7 ns |  1.13
bm_has_single_bit_if<uint32_t> | 14.7 ns  | 14.3 ns |  1.03
bm_has_single_bit_if<uint64_t> | 14.7 ns  | 14.1 ns |  1.04
bm_has_single_bit<uint8_t>     | 9.09 ns  | 8.84 ns |  1.03
bm_has_single_bit<uint16_t>    | 9.09 ns  | 7.24 ns |  1.26
bm_has_single_bit<uint32_t>    | 9.07 ns  | 7.25 ns |  1.25
bm_has_single_bit<uint64_t>    | 9.08 ns  | 7.27 ns |  1.25