@StephanTLavavej 2025-05-28
I'm mirroring this to the MSVC-internal repo - please notify me if any further changes are pushed.